requirements.txt
litellm
pyyaml
python-dotenv
rich                  # pretty terminal output